rename proposal.rs to psk.rs for now
