+++ openmls/src/group/mls_group/mod.rs
@@ -38,7 +38,7 @@ pub(crate) mod config;
 pub(crate) mod errors;
 pub(crate) mod membership;
 pub(crate) mod processing;
-pub(crate) mod proposal;
+pub(crate) mod psk;
 pub(crate) mod ser;
 
 // Tests
